+++ src/Symfony/Component/DependencyInjection/Compiler/RegisterEnvVarProcessorsPass.php
@@ -11,13 +11,11 @@
 
 namespace Symfony\Component\DependencyInjection\Compiler;
 
-use Symfony\Component\DependencyInjection\Argument\ServiceClosureArgument;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\EnvVarProcessor;
 use Symfony\Component\DependencyInjection\EnvVarProcessorInterface;
 use Symfony\Component\DependencyInjection\Exception\InvalidArgumentException;
 use Symfony\Component\DependencyInjection\ParameterBag\EnvPlaceholderParameterBag;
-use Symfony\Component\DependencyInjection\ServiceLocator;
 use Symfony\Component\DependencyInjection\Reference;
 
 /**
@@ -41,7 +39,7 @@ class RegisterEnvVarProcessorsPass implements CompilerPassInterface
                 throw new InvalidArgumentException(sprintf('Service "%s" must implement interface "%s".', $id, EnvVarProcessorInterface::class));
             }
             foreach ($class::getProvidedTypes() as $prefix => $type) {
-                $processors[$prefix] = new ServiceClosureArgument(new Reference($id));
+                $processors[$prefix] = new Reference($id);
                 $types[$prefix] = self::validateProvidedTypes($type, $class);
             }
         }
@@ -56,9 +54,8 @@ class RegisterEnvVarProcessorsPass implements CompilerPassInterface
         }
 
         if ($processors) {
-            $container->register('container.env_var_processors_locator', ServiceLocator::class)
+            $container->setAlias('container.env_var_processors_locator', (string) ServiceLocatorTagPass::register($container, $processors))
                 ->setPublic(true)
-                ->setArguments(array($processors))
             ;
         }
     }