+++ src/Symfony/Component/DependencyInjection/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 -----
 
  * added `ServiceSubscriberTrait`
+ * added `ServiceLocatorArgument` for creating optimized service-locators
 
 4.1.0
 -----